rental-backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.3
mysql-connector-python==8.0.33
Werkzeug==2.3.7
gunicorn==21.2.0